fix(rptun-pm): relax power management in rptun_dev_stop
In some cases, when `rptun_dev_stop` is called, the `rpmsg_device_destroy`
function sends an `RPMSG_NS_DESTROY` message to the peer but then immediately
shuts down the remote processor (rproc). This immediate shutdown prevents
the `rptun_pm_action` function (within the `rptun_callback`) from being
effective and relaxing the power management (pm) state.

To resolve this, we now check the pm state and relax it appropriately
before shutting down the rproc. This ensures that the pm state is properly
managed even when the remote processor is stopped abruptly.

Change-Id: Iff71953230081230a9212b100d73f0166b1dbf23